test: migrate to vitest
