+++ test/benchmark/package.json
@@ -3,7 +3,7 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "test": "node --test specs/ && echo '1'",
+    "test": "vitest run",
     "bench:json": "vitest bench --reporter=json",
     "bench": "vitest bench"
   },
